finalized endpoints

+++ InstaTonneApis/endpoints/followers.py
@@ -45,8 +45,6 @@ def author_follower_foreign(request : HttpRequest, id : str,foreign_id : str):
 
 def delete_author_follower(request : HttpRequest,id : str, foreign_id : str):
 
-    follows = Follow.objects.all().filter(followeeAuthorId=id,followerAuthorId=foreign_id)
-
     if not request.user.is_authenticated:
         #not authenticated
         return HttpResponse(status=403)
@@ -56,10 +54,20 @@ def delete_author_follower(request : HttpRequest,id : str, foreign_id : str):
         #requesting to delete followers from someone that is not you. no good
         return HttpResponse(status=403)
 
+    followee = Author.objects.filter(userID=id)
+    follower = Author.objects.filter(userID=foreign_id)
+
+    #make sure both users exist
+    if len(followee) == 0 or len(follower) == 0:
+        return HttpResponse(status=404)
+    
+    #get follows relation and make sure it exists
+    follows = Follow.objects.all().filter(followeeAuthorId=followee[0],followerAuthorId=follower[0])
+
+    print(len(follows))
     if len(follows) != 1:
         #foreign_id follower does not exist!
         return HttpResponse(status=404)
-    
     follows.delete()
 
     return HttpResponse(status=204)
@@ -70,9 +78,10 @@ def put_author_follower(request : HttpRequest,id : str, foreign_id : str):
         #not authenticated. woopsies!
         return HttpResponse(status=403)
     
-    
-    if str(request.user.pk) != id:
+    print(request.user.pk)
+    if str(request.user.pk) != foreign_id:
         #requesting someone thats not you. no bueno
+        print("test")
         return HttpResponse(status=403)
     
     if id == foreign_id:
@@ -84,7 +93,7 @@ def put_author_follower(request : HttpRequest,id : str, foreign_id : str):
     follower = Author.objects.filter(userID=foreign_id)
 
     if len(followee) != 1:
-        return HttpResponse(status=400,content="Author does not exist.")
+        return HttpResponse(status=404,content="Author does not exist.")
     
     new_follow = Follow.objects.create(followeeAuthorId=followee[0],followerAuthorId=follower[0])
